+++ controller/src/state.h
@@ -24,7 +24,7 @@ public:
 
     // Sensor results
     float speed;
-    float turningAngle;
+    int8_t turningAngle;
 
     // Environment
     bool inDarkness;
